+++ app/src/main/java/com/example/nework/repository/PostRepository.kt
@@ -6,7 +6,7 @@ import com.example.nework.dto.Post
 
 interface PostRepository : Repository {
     suspend fun save(post: Post, upload: MediaUpload?)
-    suspend fun likeById(id: Int)
-    suspend fun dislikeById(id: Int)
+    suspend fun likeById(id: Long)
+    suspend fun dislikeById(id: Long)
     suspend fun upload(upload: MediaUpload): Media
 }
